refactor(shell): Phase 2——project-overview.vue 再抽出三个方法组模块
延续 Phase 0（样式外置）/ Phase 1（三个方法组），继续拆 script：

- stagingArea.js（178 行，8 个方法）：文件暂存区
- tabDragSplit.js（296 行，14 个方法）：标签页拖拽与分栏布局
- fileOpenTabs.js（362 行，16 个方法）：文件打开与标签页生命周期

project-overview.vue 6286 → 5458 行（累计从 10639 降到 5458）。

纯搬移，零逻辑改写：三段均为连续行区间整体搬出，与原文件逐行比对
836 行全部字节一致；对象展开进 methods，this 仍指向页面实例。

顺带删掉本次改动造成的孤儿 import（createFolder / batchMoveFiles /
fileGlyph 已随方法移入模块，.vue 不再引用）。

本次刻意不动 OCR/截图簇（约 1000 行）与剪贴板组：前者分散在六段不连续
区间，无法用"整段搬移+逐行比对"证明等价；后者含 4 处 #ifdef H5 条件编译
指令，而 .js 模块里的条件编译尚无先例，留待单独验证。

验证：
- 逐行比对：836 行字节一致（脚本化搬移，非手工复制）
- 方法名扫描：六个模块 58 个方法与 .vue 剩余 156 个方法零重名
- 模板绑定扫描：模板里 35 个方法调用全部可解析
- npm run build:h5 通过；npm run check:emits 通过
- npm run test:app-e2e 连续 3 次 29/29 通过
