+++ frontend/src/pages/project-overview/project-overview.vue
@@ -1273,8 +1273,6 @@ import {
   getProjectVariables,
   getProjectFiles,
   batchCopyFiles,
-  createFolder,
-  batchMoveFiles,
   aiChat,
   exportAiDocx,
   getProjectMembers,
@@ -1305,13 +1303,16 @@ import {
 
 import { activityTracker } from '@/utils/activityTracker.js'
 
-import { ICONS as GLYPHS, fileGlyph } from '@/config/icons.js'
+import { ICONS as GLYPHS } from '@/config/icons.js'
 import DdFilesPanel from '@/components/DdFilesPanel.vue'
 import DdRequestEditor from '@/components/DdRequestEditor.vue'
 import ChatInterface from '@/components/ChatInterface.vue'
 import { panelSwitchingMethods } from './panelSwitching.js'
 import { agentClientActionMethods } from './agentClientActions.js'
 import { librePoolMethods } from './librePool.js'
+import { stagingAreaMethods } from './stagingArea.js'
+import { tabDragSplitMethods } from './tabDragSplit.js'
+import { fileOpenTabsMethods } from './fileOpenTabs.js'
 
 
 export default {
@@ -2293,6 +2294,10 @@ export default {
     ...panelSwitchingMethods,
     ...agentClientActionMethods,
     ...librePoolMethods,
+    // Phase 2 外置的方法组
+    ...stagingAreaMethods,
+    ...tabDragSplitMethods,
+    ...fileOpenTabsMethods,
     // EasyVoice Integration
     async handleEasyVoiceDocRequest(callback) {
       console.log('[EasyVoice] Requesting doc text...')
@@ -2699,184 +2704,7 @@ export default {
     },
 
     // === Staging Area Methods ===
-    async ensureStagingFolder() {
-      if (this.stagingFolderId) return
-      try {
-        // Changed from .stagezone to __staging_area__ to avoid dotfile filtering
-        const folderName = '__staging_area__'
-        console.log('EnsureStaging: Fetching root files...')
-        const response = await getProjectFiles(this.projectId, null)
-
-        // Normalize response: API returns { code: 0, data: [...] } or possibly just array
-        const files = Array.isArray(response) ? response : (response?.data || [])
-        console.log('EnsureStaging: Got files count:', files.length, 'Looking for:', folderName)
-
-        let folder = files.find ? files.find(f => f.name === folderName) : null
-        console.log('EnsureStaging: Found folder?', folder ? folder.id : null)
-
-        if (!folder) {
-          console.log('EnsureStaging: Creating folder...')
-          try {
-             const createResp = await createFolder(this.projectId, null, folderName)
-             // createFolder may also return wrapped response
-             folder = createResp?.data || createResp
-          } catch(err) {
-             console.error('EnsureStaging: Create failed:', err)
-
-             // Retry: Maybe it exists now (concurrency) or previous fetch missed it?
-             // Or maybe create failed because it already exists.
-             console.log('EnsureStaging: Retrying fetch...')
-             const retryResponse = await getProjectFiles(this.projectId, null)
-             const retryFiles = Array.isArray(retryResponse) ? retryResponse : (retryResponse?.data || [])
-             folder = retryFiles.find ? retryFiles.find(f => f.name === folderName) : null
-          }
-        }
-
-        if (folder) {
-           this.stagingFolderId = folder.id
-           console.log('EnsureStaging: stagingFolderId set to:', folder.id)
-        }
-      } catch (e) {
-        console.error('Failed to ensure staging folder:', e)
-      }
-    },
-    async loadStagingFiles() {
-      console.log('LoadStaging: Starting...')
-      if (!this.stagingFolderId) await this.ensureStagingFolder()
-      if (!this.stagingFolderId) {
-          console.warn('LoadStaging: No folder ID, aborting.')
-          return
-      }
-      try {
-        console.log('LoadStaging: Fetching files for ID:', this.stagingFolderId)
-        const response = await getProjectFiles(this.projectId, this.stagingFolderId)
-        // Normalize response: API returns { code: 0, data: [...] } or possibly just array
-        const files = Array.isArray(response) ? response : (response?.data || [])
-        console.log('LoadStaging: Got files count:', files.length)
-        this.stagingFiles = files
-        // Debug: log file names
-        if (files.length > 0) {
-          console.log('LoadStaging: Files:', files.map(f => f.name).join(', '))
-        }
-      } catch (e) {
-        console.error('Failed to load staging files:', e)
-        uni.showToast({ title: '加载暂存区文件失败', icon: 'none' })
-      }
-    },
-    async onStagingDrop(files) {
-      console.log('onStagingDrop received:', files)
-
-      // Critical Fix: Ensure files is an array.
-      // If event object passed accidentally, return.
-      if (!files) return
-
-      // Check if this is a native DOM Event object (not a Vue emitted payload)
-      // This can happen if the native @drop event bubbles up
-      if (files instanceof Event || (files.type && files.target && files.currentTarget)) {
-        // Silently ignore - the actual files should be processed by FileStagingArea's onDrop
-        return
-      }
-
-      if (!Array.isArray(files)) {
-         // Try to recover if it's a single file object
-         if (files.id && files.name) {
-             files = [files]
-         } else {
-             console.warn('onStagingDrop: Unexpected argument format, ignoring:', typeof files)
-             return
-         }
-      }
-
-      if (files.length === 0) return
-      if (!this.stagingFolderId) await this.ensureStagingFolder()
-
-      const fileIds = files.map(f => f.id)
-
-      // 记录每个文件进入暂存区前的原始 parentId
-      for (const file of files) {
-        if (file.id && file.parentId !== undefined) {
-          this.stagingOriginalParents[file.id] = file.parentId
-        }
-      }
-
-      try {
-        // Move to staging folder
-        await batchMoveFiles(this.projectId, fileIds, this.stagingFolderId)
-
-        // Reload staging files
-        await this.loadStagingFiles()
-        // Reload file tree (to remove from original location)
-        if (this.$refs.fileTree) {
-            this.$refs.fileTree.loadFiles()
-        }
-
-        // Auto-pin
-        this.stagingPinned = true
-        uni.showToast({ title: '已加入暂存区', icon: 'success' })
-      } catch (e) {
-        console.error('Failed to move files to staging:', e)
-        uni.showToast({ title: '加入暂存区失败', icon: 'none' })
-      }
-    },
-    handleStagingClear() {
-       // Optional: Move all back to root? Or just clear list (which creates orphans in .stagezone)?
-       // User didn't specify, but "Clear" usually means empty the list.
-       // Logic: Move all files in staging back to root.
-       if (this.stagingFiles.length === 0) return
-       const ids = this.stagingFiles.map(f => f.id)
-
-       batchMoveFiles(this.projectId, ids, null).then(() => {
-          this.loadStagingFiles()
-          this.$refs.fileTree.loadFiles()
-       })
-    },
-    async handleStagingRemove(id) {
-       // 恢复到文件原来的目录，如果找不到则移动到根目录
-       if (!this.stagingFolderId) return
-
-       // 获取原始 parentId（如果有记录的话）
-       const originalParentId = this.stagingOriginalParents[id]
-       // 注意：originalParentId 可能是 null（原本就在根目录）或 undefined（未记录）
-       // 两者都应该移动到根目录
-       const targetParentId = originalParentId !== undefined ? originalParentId : null
-
-       try {
-         await batchMoveFiles(this.projectId, [id], targetParentId)
-         // 清理该文件的原始目录记录
-         delete this.stagingOriginalParents[id]
-         await this.loadStagingFiles()
-         this.$refs.fileTree.loadFiles()
-       } catch (e) {
-         console.error('Failed to remove from staging:', e)
-         // 如果恢复失败（可能原目录已删除），尝试移动到根目录
-         if (targetParentId !== null) {
-           try {
-             await batchMoveFiles(this.projectId, [id], null)
-             delete this.stagingOriginalParents[id]
-             await this.loadStagingFiles()
-             this.$refs.fileTree.loadFiles()
-             uni.showToast({ title: '原目录已不存在，已移至根目录', icon: 'none' })
-             return
-           } catch (e2) {
-             console.error('Fallback to root also failed:', e2)
-           }
-         }
-         uni.showToast({ title: '移出暂存区失败', icon: 'none' })
-       }
-    },
-    handleStagingCompare(files) {
-        if (!files || files.length !== 2) return;
-        this.onCompareDocumentsRequest(files);
-    },
-    handleStagingOpen(file) {
-      if (!file) return
-      this.openFile(file)
-    },
-    handleStagingCollapse() {
-      // User explicitly collapsed staging area
-      this.stagingPinned = false
-      this.stagingManuallyCollapsed = true
-    },
+    // 文件暂存区方法组已外置 → ./stagingArea.js（Phase 2）
 
     async openFileLinkTarget(fileId, sideOverride = null) {
       const fid = Number(fileId)
@@ -4512,666 +4340,10 @@ export default {
     },
 
     // ================= Tabs 拖拽：同窗格换序 + 分屏跨窗格移动 =================
-    onTabDragStart(evt, file, fromPane) {
-      this.draggingTab = { fileId: file.id, fromPane }
-      this.tabDragOver = null
-      try {
-        if (evt && evt.dataTransfer) {
-          evt.dataTransfer.effectAllowed = 'move'
-          evt.dataTransfer.setData('application/json', JSON.stringify({ fileId: file.id, fromPane }))
-        }
-      } catch (e) {
-        // ignore
-      }
-    },
-
-    onTabDragOver(_evt, file, pane) {
-      this.tabDragOver = { fileId: file.id, pane }
-    },
-
-    onTabDropZoneDragOver(pane) {
-      // 空白区域的 hover 提示（只记录 pane，避免 tab-drag-over 误高亮）
-      if (!this.draggingTab) return
-      this.tabDragOver = null
-      // 预激活：拖到右侧区域时，先聚焦右侧，避免“未激活窗格无法接收拖拽”的体验
-      if (pane === 'right' && this.splitMode) {
-        this.focusedPane = 'right'
-      }
-    },
-
-    onTabDropOnZone(evt, targetPane) {
-      const payload = this.getTabDragPayload(evt) || this.draggingTab
-      if (!payload || !payload.fileId) return
-      // drop 到空白区域：插入到末尾
-      this.moveTabTo(payload.fileId, payload.fromPane, targetPane, null)
-      this.onTabDragEnd()
-    },
-
-    onTabDropOnItem(evt, targetFile, targetPane) {
-      const payload = this.getTabDragPayload(evt) || this.draggingTab
-      if (!payload || !payload.fileId) return
-
-      this.moveTabTo(payload.fileId, payload.fromPane, targetPane, targetFile.id)
-      this.onTabDragEnd()
-    },
-
-    getTabDragPayload(evt) {
-      try {
-        const raw = evt?.dataTransfer?.getData('application/json')
-        if (!raw) return null
-        return JSON.parse(raw)
-      } catch (e) {
-        return null
-      }
-    },
-
-    moveTabTo(fileId, fromPane, toPane, beforeFileId) {
-      if (!fileId || !fromPane || !toPane) return
-      if (!this.splitMode && toPane === 'right') return
-
-      const fromList = fromPane === 'left' ? this.leftFiles : this.rightFiles
-      const toList = toPane === 'left' ? this.leftFiles : this.rightFiles
-
-      const fromIdx = fromList.findIndex(f => f.id === fileId)
-      if (fromIdx < 0) return
-
-      const source = fromList[fromIdx]
-      // 关键修复：跨窗格拖拽不“移动”，而是“在另一侧打开同一文件”（允许左右双开）
-      const isCrossPane = fromPane !== toPane
-      const moved = isCrossPane ? { ...source } : fromList.splice(fromIdx, 1)[0]
-
-      // 目标索引：插入到 beforeFileId 前面（如果没有则追加末尾）
-      let toIdx = -1
-      if (beforeFileId) {
-        toIdx = toList.findIndex(f => f.id === beforeFileId)
-      }
-      // 若目标窗格已存在同文件，则仅激活，不重复插入
-      const existedIdx = toList.findIndex(f => f.id === moved.id)
-      if (existedIdx >= 0) {
-        // 如果目标还指定了 beforeFileId 且是同窗格换序，可以做排序调整
-        if (!isCrossPane && beforeFileId && existedIdx !== toIdx && toIdx >= 0) {
-          const [existing] = toList.splice(existedIdx, 1)
-          toList.splice(toIdx, 0, existing)
-        }
-      } else {
-        if (toIdx < 0) {
-          toList.push(moved)
-        } else {
-          toList.splice(toIdx, 0, moved)
-        }
-      }
-
-      // 激活态跟随：如果是跨窗格移动，更新 activeFileId
-      if (toPane === 'left') {
-        this.activeFileIdLeft = moved.id
-        this.focusedPane = 'left'
-      } else {
-        this.activeFileIdRight = moved.id
-        this.focusedPane = 'right'
-      }
-
-      // 拖拽换序/跨窗格会改变容器尺寸分配，给 WPS 一个 resize
-      this.$nextTick(() => this.triggerWorkbenchResize())
-    },
-
-    onTabDragEnd() {
-      this.draggingTab = null
-      this.tabDragOver = null
-    },
-
-    isOpenInOtherPane(fileId, pane) {
-      if (!fileId) return false
-      if (pane === 'left') return this.rightFiles.some(f => f.id === fileId)
-      return this.leftFiles.some(f => f.id === fileId)
-    },
-
-    startResize(target, evt) {
-      // target: 'left' | 'right' | 'bottom'
-      this.resizing.active = true
-      this.resizing.target = target
-      const e = evt && evt.touches && evt.touches[0] ? evt.touches[0] : evt
-      this.resizing.startX = e?.clientX || 0
-      this.resizing.startY = e?.clientY || 0
-      this.resizing.startSidebarWidth = this.sidebarWidth
-      this.resizing.startAiWidth = this.aiPanelWidth
-      this.resizing.startToolsHeight = this.toolsPanelHeight
-
-      // Cache DOM element for direct manipulation
-      this.resizing.element = null
-      if (target === 'left') {
-        this.resizing.element = this.$refs.sidebarLeft ? (this.$refs.sidebarLeft.$el || this.$refs.sidebarLeft) : null
-      } else if (target === 'right') {
-        this.resizing.element = this.$refs.aiPanel ? (this.$refs.aiPanel.$el || this.$refs.aiPanel) : null
-      } else if (target === 'bottom') {
-         this.resizing.element = this.$refs.bottomPanel ? (this.$refs.bottomPanel.$el || this.$refs.bottomPanel) : null
-      }
-
-      if (evt && typeof evt.preventDefault === 'function') {
-        evt.preventDefault()
-      }
-
-      if (typeof window !== 'undefined') {
-        if (!this.boundResizeMove) this.boundResizeMove = (e2) => this.onResizeMove(e2)
-        if (!this.boundStopResize) this.boundStopResize = () => this.stopResize()
-
-        window.addEventListener('mousemove', this.boundResizeMove, { passive: false })
-        window.addEventListener('mouseup', this.boundStopResize, { passive: true })
-        window.addEventListener('touchmove', this.boundResizeMove, { passive: false })
-        window.addEventListener('touchend', this.boundStopResize, { passive: true })
-        // 兜底：鼠标移出窗口/窗口失焦时 mouseup 可能丢失，避免 is-resizing 卡死
-        window.addEventListener('blur', this.boundStopResize, { passive: true })
-      }
-    },
-
-    onResizeMove(evt) {
-      if (!this.resizing.active) return
-      const e = evt && evt.touches && evt.touches[0] ? evt.touches[0] : evt
-      const clientX = e?.clientX || 0
-      const clientY = e?.clientY || 0
-
-      // 防止页面滚动
-      if (evt && typeof evt.preventDefault === 'function') {
-        evt.preventDefault()
-      }
-
-      // rAF 节流：让拖拽跟手更稳（避免 move 事件过密导致抖动/延迟）
-      this._resizePendingX = clientX
-      this._resizePendingY = clientY
-      if (this._resizeRaf) return
-      this._resizeRaf = requestAnimationFrame(() => {
-        this._resizeRaf = null
-        const vw = typeof window !== 'undefined' ? window.innerWidth : 1200
-        const vh = typeof window !== 'undefined' ? window.innerHeight : 800
-
-        // Cursor 体验：拖动范围尽量大，不强行保证中间工作区可用（遮挡就遮挡）
-        const leftMin = 160
-        const leftMax = Math.max(leftMin, Math.floor(vw * 0.75))
-        const rightMin = 240
-        const rightMax = Math.max(rightMin, Math.floor(vw * 0.75))
-        const headerH = 56
-        const tabsH = 40
-        const bottomMin = 140
-        const bottomMax = Math.max(bottomMin, Math.floor((vh - headerH - tabsH) * 0.85))
-
-        const dx = (this._resizePendingX || 0) - this.resizing.startX
-        const dy = (this._resizePendingY || 0) - this.resizing.startY
-
-        let finalValue = 0
-
-        if (this.resizing.target === 'left') {
-          const next = this.resizing.startSidebarWidth + dx
-          finalValue = Math.max(leftMin, Math.min(leftMax, next))
-          // Direct DOM update
-          if (this.resizing.element) {
-              this.resizing.element.style.width = finalValue + 'px'
-          }
-          // Store for sync on stop
-          this.resizing.currentValue = finalValue
-        } else if (this.resizing.target === 'right') {
-          const next = this.resizing.startAiWidth - dx
-          finalValue = Math.max(rightMin, Math.min(rightMax, next))
-           // Direct DOM update
-           if (this.resizing.element) {
-              this.resizing.element.style.width = finalValue + 'px'
-          }
-          this.resizing.currentValue = finalValue
-        } else if (this.resizing.target === 'bottom') {
-          const next = this.resizing.startToolsHeight - dy
-          finalValue = Math.max(bottomMin, Math.min(bottomMax, next))
-           // Direct DOM update
-           if (this.resizing.element) {
-              this.resizing.element.style.height = finalValue + 'px'
-          }
-           this.resizing.currentValue = finalValue
-        }
-      })
-    },
-
-    stopResize() {
-      if (!this.resizing.active) return
-
-      // Save target and currentValue BEFORE nullifying
-      const target = this.resizing.target
-      const finalValue = this.resizing.currentValue
-
-      this.resizing.active = false
-      this.resizing.target = null
-      this.resizing.element = null
-      this.resizing.currentValue = null
-
-      if (this._resizeRaf) {
-        cancelAnimationFrame(this._resizeRaf)
-        this._resizeRaf = null
-      }
-
-      if (typeof window !== 'undefined') {
-        if (this.boundResizeMove) {
-          window.removeEventListener('mousemove', this.boundResizeMove)
-          window.removeEventListener('touchmove', this.boundResizeMove)
-        }
-        if (this.boundStopResize) {
-          window.removeEventListener('mouseup', this.boundStopResize)
-          window.removeEventListener('touchend', this.boundStopResize)
-          window.removeEventListener('blur', this.boundStopResize)
-        }
-      }
-
-      // Sync final value to Vue state using saved target
-      if (finalValue) {
-          if (target === 'left') {
-              this.sidebarWidth = finalValue
-          } else if (target === 'right') {
-              this.aiPanelWidth = finalValue
-          } else if (target === 'bottom') {
-              this.toolsPanelHeight = finalValue
-          }
-      }
-
-      this.$nextTick(() => this.triggerWorkbenchResize())
-    },
-
-    toggleSplitMode() {
-      this.splitMode = !this.splitMode
-
-      // 关键修复：触发 resize 事件通知 WPS SDK 调整布局
-      // WPS SDK 监听 window resize 来调整内部 iframe 大小
-      this.$nextTick(() => this.triggerWorkbenchResize())
-
-      if (!this.splitMode) {
-        // 关闭分屏时，重置 focus 到左侧
-        this.focusedPane = 'left'
-      } else {
-        // 开启分屏时，默认聚焦右侧，方便用户立即选择文件
-        this.focusedPane = 'right'
-      }
-    },
-    focusPane(pane) {
-      // 只有在分屏模式下才允许聚焦右侧
-      if (!this.splitMode && pane === 'right') return
-
-      const oldPane = this.focusedPane
-      this.focusedPane = pane
-
-      if (oldPane !== pane) {
-          // Switch active tracking to the file in the new pane
-          const file = pane === 'left' ? this.activeFileLeft : this.activeFileRight
-          if (file) {
-              const meta = this.project && this.project.name ? `Project: ${this.project.name}` : ''
-              if (this.isBrowserTab(file)) {
-                   const url = file.url || ''
-                   const title = file.name || ''
-                   const fullMeta = meta + (title ? `. Title: ${title}` : '')
-                   activityTracker.trackActivePage('OPEN_URL', 0, url, fullMeta)
-              } else {
-                   activityTracker.trackActivePage('OPEN_FILE', file.id, file.name, meta)
-              }
-          }
-      }
-    },
+    // 标签页拖拽与分栏方法组已外置 → ./tabDragSplit.js（Phase 2）
 
     // --- 文件管理逻辑 ---
-    handleFileTreeSelect(file) {
-      if (!file || file.isFolder) return
-      this.openFile(file)
-    },
-
-    // Handle file open from SearchPanel
-    handleSearchOpenFile(file) {
-      if (!file) return
-      console.log('[project-overview] Open file from search:', file)
-      this.openFile({
-        id: file.id,
-        wpsFileId: file.wpsFileId,
-        name: file.name,
-        fileType: file.fileType,
-        filePath: file.filePath
-      })
-    },
-
-    // Handle file open request from ChatInterface (file changes popup)
-    async handleOpenFileFromChat({ name }) {
-      if (!name) return
-      console.log('[project-overview] Open file from chat:', name)
-
-      // Refresh project files first to ensure we have latest
-      try {
-        const resp = await getProjectFiles(this.projectId)
-        // Normalize response: API returns { code: 0, data: [...] } or possibly just array
-        const files = Array.isArray(resp) ? resp : (resp?.data || [])
-        console.log('[project-overview] Got files for search:', files.length)
-
-        // Find file by name (case-insensitive, match basename)
-        const targetFile = files.find(f => {
-          if (f.isFolder) return false
-          // Match exact name or name without extension
-          return f.name === name || f.name.toLowerCase() === name.toLowerCase()
-        })
-
-        if (targetFile) {
-          console.log('[project-overview] Found file:', targetFile.id, targetFile.name)
-          this.openFile(targetFile)
-        } else {
-          console.warn('File not found:', name, 'in', files.map(f => f.name))
-          uni.showToast({ title: '未找到文件: ' + name, icon: 'none' })
-        }
-      } catch (e) {
-        console.error('Failed to fetch files for open:', e)
-        uni.showToast({ title: '获取文件列表失败', icon: 'none' })
-      }
-    },
-
-    openFile(file) {
-      // 检查文件类型是否支持打开
-      if (!this.isFileTypeSupported(file)) {
-        uni.showModal({
-          title: '无法打开文件',
-          content: `暂不支持打开此类型文件：${file.name}\n\n文件类型：${file.fileType || '无后缀名'}\n\n支持的文件类型：\n• 文档：doc, docx, xls, xlsx, ppt, pptx, pdf\n• 图片：jpg, jpeg, png, gif, bmp, webp, svg\n• 视频：mp4, webm, ogg, mov, mkv, avi\n• 音频：mp3, wav, m4a, flac, aac\n• 文本：txt, md, json, xml, html等`,
-          showCancel: false,
-          confirmText: '我知道了',
-          success: (res) => {
-            if (res.confirm) {
-              console.log('用户确认无法打开文件')
-            }
-          }
-        })
-        return
-      }
-
-      const meta = this.project && this.project.name ? `Project: ${this.project.name}` : ''
-      // Start session tracking for this file
-      activityTracker.trackActivePage('OPEN_FILE', file.id, file.name, meta)
-
-      // 1. 如果已经在某个 pane 打开，则聚焦该 pane
-      const existingLeft = this.leftFiles.find(f => f.id === file.id)
-      // - 如果当前聚焦窗格未打开该文件，则在当前窗格打开
-      // - 若当前窗格已打开，则仅激活
-      const targetPane = this.splitMode ? this.focusedPane : 'left'
-      const targetList = targetPane === 'left' ? this.leftFiles : this.rightFiles
-      const targetIdProp = targetPane === 'left' ? 'activeFileIdLeft' : 'activeFileIdRight'
-
-      const existing = targetList.find(f => f.id === file.id)
-      if (existing) {
-        Object.assign(existing, file)
-        this[targetIdProp] = file.id
-        this.focusedPane = targetPane
-      } else {
-        targetList.push({ ...file })
-        this[targetIdProp] = file.id
-      }
-
-      // Persist active ID for current mode
-      const mode = this.leftPaneKey || 'files'
-      this.lastActiveIdsByMode[targetPane][mode] = file.id
-      this.saveActiveIdsByMode()
-
-      // 打开/激活后，给 WPS 一个机会刷新（避免容器尺寸/激活状态不对）
-      this.$nextTick(() => this.triggerWorkbenchResize())
-    },
-
-    activateTab(file, pane) {
-      // 点击 Tab 时，切换对应窗格的激活文件，并聚焦该窗格
-      this.focusPane(pane)
-      if (pane === 'left') {
-        this.activeFileIdLeft = file.id
-      } else {
-        this.activeFileIdRight = file.id
-      }
-
-      // Persist active ID for current mode
-      const mode = this.leftPaneKey || 'files'
-      this.lastActiveIdsByMode[pane][mode] = file.id
-      this.saveActiveIdsByMode()
-
-      // 同步更新 FileTree 的选中状态
-      // 如果是文件类型（非浏览器标签、非特殊标签类型），需要更新资源管理器的选中状态
-      if (!this.isBrowserTab(file) && file.id && !file.tabType) {
-        // 展开侧边栏并切换到文件模式
-        if (this.sidebarCollapsed) {
-          this.sidebarCollapsed = false
-        }
-        if (this.leftPaneKey !== 'files') {
-          this.leftPaneKey = 'files'
-        }
-
-        // 确保在下一个 tick 中执行，此时 FileTree 组件已经更新
-        this.$nextTick(() => {
-          if (this.$refs.fileTree) {
-            // 使用 revealFile 方法来定位并选中文件（会展开父目录并滚动到文件）
-            if (this.$refs.fileTree.revealFile) {
-              this.$refs.fileTree.revealFile(file.id)
-            } else {
-              // 如果没有 revealFile 方法，直接设置 selectedFileId
-              this.$refs.fileTree.selectedFileId = file.id
-            }
-          }
-        })
-      } else {
-        // 如果是浏览器标签或其他特殊类型，清空 FileTree 的选中状态
-        this.$nextTick(() => {
-          if (this.$refs.fileTree) {
-            this.$refs.fileTree.selectedFileId = null
-            this.$refs.fileTree.multiSelectedIds = []
-          }
-        })
-      }
-
-      // Track switch
-      const meta = this.project && this.project.name ? `Project: ${this.project.name}` : ''
-      if (this.isBrowserTab(file)) {
-          // If browser tab, we need to track URL session
-          const url = file.url || ''
-          const title = file.name || ''
-          const fullMeta = meta + (title ? `. Title: ${title}` : '')
-          activityTracker.trackActivePage('OPEN_URL', 0, url, fullMeta)
-      } else {
-          // File
-          activityTracker.trackActivePage('OPEN_FILE', file.id, file.name, meta)
-      }
-    },
-
-    handleFileDeleted(payload) {
-      if (!payload || !payload.ids || !Array.isArray(payload.ids)) return
-      const ids = new Set(payload.ids)
-
-      // Close tabs if they match deleted IDs
-      // Iterate backwards to avoid index issues when closing (though closeFile uses findIndex, safety first)
-      
-      // Right Pane
-      for (let i = this.rightFiles.length - 1; i >= 0; i--) {
-        if (ids.has(this.rightFiles[i].id)) {
-           this.closeFile(this.rightFiles[i].id, 'right')
-        }
-      }
-      
-      // Left Pane
-      for (let i = this.leftFiles.length - 1; i >= 0; i--) {
-        if (ids.has(this.leftFiles[i].id)) {
-           this.closeFile(this.leftFiles[i].id, 'left')
-        }
-      }
-    },
-
-    async closeFile(fileId, pane) {
-      const list = pane === 'left' ? this.leftFiles : this.rightFiles
-      const idProp = pane === 'left' ? 'activeFileIdLeft' : 'activeFileIdRight'
-
-      let idx = list.findIndex(f => f.id === fileId)
-      if (idx === -1) return
-
-      const file = list[idx]
-
-      // (autosave) Office 文档出池即卸载，卸载后 webview 没了就没法导出——
-      // 有未保存修改（或保存在途）的实例先落盘再关。加载失败的实例跳过
-      // （画布是空白原型，保存会覆盖真文件，同 evictLibreInstance）。
-      if (file && this.useLibreEditor(file)) {
-        const inst = (this._libreRefs || {})[pane + ':' + fileId]
-        if (inst && inst.ready && !inst.isError && inst.file && (inst.dirty || inst.saving)) {
-          try { await inst.flushSave() } catch (e) { console.warn('[ProjectOverview] close flush-save failed:', e) }
-        }
-        // 落盘期间列表可能已变（并发关闭）——重新定位，已被移除则到此为止
-        idx = list.findIndex(f => f.id === fileId)
-        if (idx === -1) return
-      }
-      const activeId = this[idProp]
-
-      // If closing the active file/tab, the activityTracker.trackActivePage in activateTab or openFile will handle the switch.
-      // But if we close the *currently active* file and no other file becomes active (e.g. empty list), we should stop session?
-      // Actually, if we close active file, we usually switch to another one (logic below).
-      // So we don't need to manually stop session here, UNLESS the list becomes empty.
-
-      // const meta = this.project && this.project.name ? `Project: ${this.project.name}` : ''
-      // activityTracker.logAction('CLOSE_FILE', file.id, file.name, 0, meta)
-
-      list.splice(idx, 1)
-
-      // 如果关闭的是当前激活的文件，尝试切换到临近的文件
-      if (activeId === fileId) {
-        const newActiveId = list.length > 0
-          ? list[Math.min(idx, list.length - 1)].id
-          : null
-        this[idProp] = newActiveId
-
-        // Update persisted record
-        const mode = this.leftPaneKey || 'files'
-        this.lastActiveIdsByMode[pane][mode] = newActiveId
-        this.saveActiveIdsByMode()
-      }
-    },
-
-    /** 标签页图标的 SVG path 集合（界面禁用 emoji，图标一律 stroke 线性 SVG） */
-    getFileIconPaths(type, tabType) {
-      return tabType === 'web' ? GLYPHS.web : fileGlyph(type)
-    },
-    isFileTypeSupported(file) {
-      if (!file || file.isFolder) return true
-      if (!file.fileType) return false
-
-      const type = file.fileType.toLowerCase()
-
-      // 支持的文件类型列表
-      const supportedTypes = [
-        // Office文档
-        'doc', 'docx', 'xls', 'xlsx', 'ppt', 'pptx', 'pdf',
-        // 图片
-        'jpg', 'jpeg', 'png', 'gif', 'bmp', 'svg', 'webp',
-        // 视频
-        'mp4', 'webm', 'ogg', 'mov', 'mkv', 'avi',
-        // 音频
-        'mp3', 'wav', 'm4a', 'flac', 'aac',
-        // 文本文件
-        'txt', 'md', 'markdown', 'json', 'xml', 'html', 'css', 'js', 'java', 'py', 'sh', 'sql', 'log',
-        // 压缩包（FilePreview 条目预览 + 解压）
-        'zip', 'rar', '7z',
-        // 尽调清单
-        'dd'
-      ]
-
-      return supportedTypes.includes(type)
-    },
-    isEditorOpenableFile(file) {
-      // 判断是否为「文档编辑器可打开」的 Office 类文件
-      if (!file || file.tabType === 'web' || file.tabType === 'markdown' || !file.fileType) return false
-
-      const type = file.fileType.toLowerCase()
-
-      // 1. Force native preview for media types (Images, Video, Audio)
-      const mediaTypes = [
-          // Images
-          'jpg','jpeg','png','gif','bmp','svg','webp',
-          // Video
-          'mp4','webm','ogg','mov','mkv','avi',
-          // Audio
-          'mp3','wav','m4a','flac','aac'
-      ]
-      if (mediaTypes.includes(type)) return false
-
-      // 2. 排除 Markdown 文件，使用专门的 Markdown 预览组件
-      if (type === 'md' || type === 'markdown') return false
-
-      // 3. Office 文档格式 —— 仅限自建 LOWA 引擎真正编入的模块（probe_modules
-      // 实测：仅 Writer + Calc；Impress/Draw 已裁）。
-      // - Presentation 一律走 FilePreview（pptx = pptx-preview 前端渲染）；
-      // - PDF 走 FilePreview 的 Chromium 原生渲染——LOWA 无 Draw 时 Writer 会把
-      //   PDF 二进制当文本导入，满屏乱码（2026-07 真机截图证实）。
-      const wpsFormats = [
-          // Writer
-          'wps', 'wpt', 'doc', 'dot', 'docx', 'dotx', 'docm', 'dotm', 'rtf', 'odt',
-          // Spreadsheet
-          'et', 'ett', 'ets', 'xls', 'xlsx', 'xlt', 'xltx', 'xlsm', 'xltm', 'xlsb', 'csv'
-      ]
-
-      // Office 类型或带文件 ID（非媒体/markdown）即视为文档编辑器可打开
-      return wpsFormats.includes(type) || (file.wpsFileId && !mediaTypes.includes(type) && type !== 'md' && type !== 'markdown')
-    },
-
-    // Epic #43 Track B / #79: should this Office file open in the embedded
-    // LibreOffice editor? True only when the desktop embed is available
-    // (libreOfficePreferred). On web/h5 or when unavailable, returns false and
-    // the file falls through to FilePreview (docx 本地只读渲染).
-    useLibreEditor(file) {
-      return this.libreOfficePreferred && this.isEditorOpenableFile(file)
-    },
-
-    // Check if file is a markdown tab (for AI artifacts or real .md files)
-    isMarkdownTab(file) {
-      if (!file) return false
-      // 1. AI 创建的虚拟 markdown 标签
-      if (file.tabType === 'markdown') return true
-      // 2. 真正的 .md 文件（从文件树打开）
-      if (file.fileType && (file.fileType.toLowerCase() === 'md' || file.fileType.toLowerCase() === 'markdown')) return true
-      return false
-    },
-
-    isDiffTab(file) {
-      return file && file.tabType === 'diff'
-    },
-
-    // --- 文档对比逻辑 ---
-    onCompareDocumentsRequest(docs) {
-      // FileTree 发起的文档对比请求
-      if (!docs || docs.length !== 2) {
-        uni.showToast({ title: '请选择两个文档进行对比', icon: 'none' })
-        return
-      }
-      this.compareDocuments = docs
-      this.showCompareDialog = true
-    },
-
-    onCompareDialogConfirm({ source, target }) {
-      // 用户确认了源文档和目标文档，打开 diff 标签页
-      this.showCompareDialog = false
-      this.openDiffTab(source, target)
-    },
-
-    openDiffTab(source, target) {
-      // 创建 diff 类型的虚拟标签页
-      const diffId = `diff-${source.id}-${target.id}-${Date.now()}`
-      const diffFile = {
-        id: diffId,
-        name: `${source.name} ↔ ${target.name}`,
-        tabType: 'diff',
-        fileType: 'diff',
-        diffSource: {
-          id: source.id,
-          name: source.name
-        },
-        diffTarget: {
-          id: target.id,
-          name: target.name
-        },
-        createdAt: Date.now()
-      }
-
-      // 添加到当前窗格
-      const targetPane = this.splitMode ? this.focusedPane : 'left'
-      const targetList = targetPane === 'left' ? this.leftFiles : this.rightFiles
-      const targetIdProp = targetPane === 'left' ? 'activeFileIdLeft' : 'activeFileIdRight'
-
-      targetList.push(diffFile)
-      this[targetIdProp] = diffFile.id
-
-      console.log('[ProjectOverview] 打开文档对比标签:', diffFile.name)
-    },
+    // 文件打开与标签页方法组已外置 → ./fileOpenTabs.js（Phase 2）
 
     // 处理文件重命名（历史：WPS 时代由文件信息轮询触发；现无调用方，保留为通用逻辑）
     async handleFileRename(pane, data) {
